rename: pipeline.ts -> mandible.ts
