+++ src/dsl/index.ts
@@ -1,2 +1,2 @@
 export { colony, ColonyBuilder } from './builder.js';
-export { mandible, pipeline, MandibleBuilder } from './pipeline.js';
+export { mandible, pipeline, MandibleBuilder } from './mandible.js';
